Reorg to lib and separate test directory.
