+++ src/lib.rs
@@ -1,2 +1,4 @@
 pub mod api;
 pub mod telemetry;
+pub mod schema;
+pub mod lrgs;
