+++ src/lrgs/controller.rs
@@ -1,11 +1,6 @@
 use std::{sync::Arc, time::Duration};
 
-use crate::{
-    config::{create_lrgs_config, create_managed_users},
-    configmap::created_script_config_map,
-    service::create_service,
-    statefulset::create_statefulset,
-};
+
 use anyhow::anyhow;
 use chrono::Utc;
 use futures::StreamExt;
@@ -18,15 +13,14 @@ use kube::{
     runtime::{controller::Action, reflector::ObjectRef, watcher, Controller},
     Api, Client, Error, Resource, ResourceExt,
 };
-use opendcs_controllers::{
+use crate::{
     api::v1::{
         dds_recv::DdsConnection,
         lrgs::{LrgsCluster, LrgsClusterStatus},
-    },
-    telemetry::{
+    }, lrgs::{config::{create_lrgs_config, create_managed_users}, configmap::created_script_config_map, service::create_service, statefulset::create_statefulset}, telemetry::{
         state::{Context, State},
         telemetry,
-    },
+    }
 };
 use serde_json::json;
 use tracing::{error, field, info, instrument, warn, Span};
