+++ src/lrgs/mod.rs
@@ -4,6 +4,3 @@ pub mod statefulset;
 pub mod configmap;
 pub mod service;
 pub mod controller;
-pub mod state;
-
-pub mod main;
