+++ src/schema/mod.rs
@@ -1,0 +1,3 @@
+pub mod job;
+pub mod configmap;
+pub mod controller;
